requirements.txt
Pillow==9.4.0
pycryptodome==3.16.0
tk==0.1.0
